perf(gdn): cast at the out_proj boundary; scope the fp16 work honestly
The GDN recurrence and its gated norm can accumulate wider than the model dtype, and
letting that reach out_proj mismatches the projection. The cast is a no-op whenever they
already agree, which is every bf16 checkpoint, and 13 GDN/qwen35moe tests confirm Ornith
is unaffected.

fp16 is still NOT working end to end and this does not claim otherwise. With the
embedding, mixer and GDN boundaries fixed the failure moves to the MoE router, with the
residual arriving float32 against fp16 weights. I stopped there rather than casting at a
fourth boundary: each fix moved the error without my having established where the float32
actually originates, and guessing at boundaries is how you end up with a model that runs
and is subtly wrong. Enabling fp16 needs a systematic dtype audit of the shared qwen35moe
modules, which is scoped work with real numerical risk to Ornith.

What is established and worth acting on later: sm_75 has no bf16 tensor cores, and on the
Quadro RTX 6000 batched bf16 GEMMs are 12-22x slower than fp16 (bs=4: 864 us vs 40 us;
prefill 300: 791 us vs 64 us), while bs=1 decode is identical. So fp16 would mainly buy
prefill and concurrency, not single-stream decode -- and it explains why concurrency 4
only returned 1.19x aggregate.

bf16 remains the serving configuration: 10.87 tok/s, all six factual completions correct.
469 tests pass; the 2 FP8 failures are pre-existing.

+++ python/freetoken/models/qwen4exp/model.py
@@ -185,8 +185,13 @@ class Qwen4ExpModel(BaseOP):
         return self.ple.gather(torch.from_numpy(rows), input_ids.device)
 
     def forward(self, input_ids: torch.Tensor) -> torch.Tensor:
-        # The wide residual starts as hc identical copies of the embedding.
+        # The wide residual starts as hc identical copies of the embedding, in the dtype
+        # the loaded weights actually have. Reading it off a real parameter here rather than
+        # from a construction-time default means the residual cannot drift into a wider
+        # dtype than the projections it feeds -- which both mismatches them and, on Turing,
+        # would pick a GEMM path 12-22x slower than the one asked for.
         state = hc_init(self.embed_tokens.forward(input_ids), self.hc)
+        state = state.to(self.hc_head.norm.dtype)
 
         # Normally the PLE embedding arrives as a graph input, gathered on the host before
         # the capture boundary (see Qwen4ExpForCausalLM.prepare_host_inputs). Warmup and
